requirements.txt
comtypes==1.1.7
lxml==4.5.0
python-docx==0.8.10
